+++ packages/backend-common/config.d.ts
@@ -29,7 +29,7 @@ export interface Config {
           /** Address of the interface that the backend should bind to. */
           address?: string;
           /** Port that the backend should listen to. */
-          port?: number;
+          port?: string | number;
         };
 
     /** HTTPS configuration for the backend. If omitted the backend will serve HTTP */
@@ -79,15 +79,24 @@ export interface Config {
       optionsSuccessStatus?: number;
     };
 
-    /**  */
-    csp?: object;
+    /**
+     * Content Security Policy options.
+     *
+     * The keys are the plain policy ID, e.g. "upgrade-insecure-requests". The
+     * values are on the format that the helmet library expects them, as an
+     * array of strings. There is also the special value false, which means to
+     * remove the default value that Backstage puts in place for that policy.
+     */
+    csp?: { [policyId: string]: string[] | false };
   };
 
   /** Configuration for integrations towards various external repository provider systems */
   integrations?: {
     /** Integration configuration for Azure */
     azure?: Array<{
-      /** The hostname of the given Azure instance */
+      /**
+       * The hostname of the given Azure instance
+       */
       host: string;
       /**
        * Token used to authenticate requests.
@@ -98,14 +107,18 @@ export interface Config {
 
     /** Integration configuration for BitBucket */
     bitbucket?: Array<{
-      /** The hostname of the given Bitbucket instance */
+      /**
+       * The hostname of the given Bitbucket instance
+       */
       host: string;
       /**
        * Token used to authenticate requests.
        * @visibility secret
        */
       token?: string;
-      /** The base url for the BitBucket API, for example https://api.bitbucket.org/2.0 */
+      /**
+       * The base url for the BitBucket API, for example https://api.bitbucket.org/2.0
+       */
       apiBaseUrl?: string;
       /**
        * The username to use for authenticated requests.
@@ -121,22 +134,30 @@ export interface Config {
 
     /** Integration configuration for GitHub */
     github?: Array<{
-      /** The hostname of the given GitHub instance */
+      /**
+       * The hostname of the given GitHub instance
+       */
       host: string;
       /**
        * Token used to authenticate requests.
        * @visibility secret
        */
       token?: string;
-      /** The base url for the GitHub API, for example https://api.github.com */
+      /**
+       * The base url for the GitHub API, for example https://api.github.com
+       */
       apiBaseUrl?: string;
-      /** The base url for GitHub raw resources, for example https://raw.githubusercontent.com */
+      /**
+       * The base url for GitHub raw resources, for example https://raw.githubusercontent.com
+       */
       rawBaseUrl?: string;
     }>;
 
     /** Integration configuration for GitLab */
     gitlab?: Array<{
-      /** The hostname of the given GitLab instance */
+      /**
+       * The hostname of the given GitLab instance
+       */
       host: string;
       /**
        * Token used to authenticate requests.