+++ packages/catalog-model/src/kinds/ComponentEntityV1alpha1.ts
@@ -30,6 +30,17 @@ const schema = yup.object<Partial<ComponentEntityV1alpha1>>({
       lifecycle: yup.string().required().min(1),
       owner: yup.string().required().min(1),
       implementsApis: yup.array(yup.string().required()).notRequired(),
+      providesApis: yup.array(yup.string().required()).notRequired(),
+      consumesApis: yup.array(yup.string().required()).notRequired(),
+      kubernetes: yup
+        .object<any>({
+          selector: yup
+            .object<any>({
+              matchLabels: yup.object<any>().required(),
+            })
+            .required(),
+        })
+        .notRequired(),
     })
     .required(),
 });
@@ -41,7 +52,21 @@ export interface ComponentEntityV1alpha1 extends Entity {
     type: string;
     lifecycle: string;
     owner: string;
+    /**
+     * @deprecated This field will disappear on Dec 14th, 2020. Please remove
+     *             any consuming code. The new field providesApis provides the
+     *             same functionality like before.
+     */
     implementsApis?: string[];
+    providesApis?: string[];
+    consumesApis?: string[];
+    kubernetes?: {
+      selector: {
+        matchLabels: {
+          [key: string]: string;
+        };
+      };
+    };
   };
 }