+++ plugins/auth-backend/src/service/router.ts
@@ -14,6 +14,14 @@
  * limitations under the License.
  */
 
+import express from 'express';
+import Router from 'express-promise-router';
+import cookieParser from 'cookie-parser';
+import { Logger } from 'winston';
+import {
+  defaultAuthProviderFactories,
+  AuthProviderFactory,
+} from '../providers';
 import {
   NotFoundError,
   PluginDatabaseManager,
@@ -21,18 +29,18 @@ import {
 } from '@backstage/backend-common';
 import { CatalogClient } from '@backstage/catalog-client';
 import { Config } from '@backstage/config';
-import cookieParser from 'cookie-parser';
-import express from 'express';
-import Router from 'express-promise-router';
-import { Logger } from 'winston';
 import { createOidcRouter, DatabaseKeyStore, TokenFactory } from '../identity';
-import { createAuthProvider } from '../providers';
+import session from 'express-session';
+import passport from 'passport';
+
+type ProviderFactories = { [s: string]: AuthProviderFactory };
 
 export interface RouterOptions {
   logger: Logger;
   database: PluginDatabaseManager;
   config: Config;
   discovery: PluginEndpointDiscovery;
+  providerFactories?: ProviderFactories;
 }
 
 export async function createRouter({
@@ -40,6 +48,7 @@ export async function createRouter({
   config,
   discovery,
   database,
+  providerFactories,
 }: RouterOptions): Promise<express.Router> {
   const router = Router();
 
@@ -59,17 +68,36 @@ export async function createRouter({
   });
   const catalogApi = new CatalogClient({ discoveryApi: discovery });
 
-  router.use(cookieParser());
+  const secret = config.getOptionalString('auth.session.secret');
+  if (secret) {
+    router.use(cookieParser(secret));
+    // TODO: Configure the server-side session storage.  The default MemoryStore is not designed for production
+    router.use(session({ secret, saveUninitialized: false, resave: false }));
+    router.use(passport.initialize());
+    router.use(passport.session());
+  } else {
+    router.use(cookieParser());
+  }
   router.use(express.urlencoded({ extended: false }));
   router.use(express.json());
 
+  const allProviderFactories = {
+    ...defaultAuthProviderFactories,
+    ...providerFactories,
+  };
   const providersConfig = config.getConfig('auth.providers');
   const providers = providersConfig.keys();
 
   for (const providerId of providers) {
     logger.info(`Configuring provider, ${providerId}`);
     try {
-      const provider = createAuthProvider(providerId, {
+      const providerFactory = allProviderFactories[providerId];
+      if (!providerFactory) {
+        throw Error(`No auth provider available for '${providerId}'`);
+      }
+
+      const provider = providerFactory({
+        providerId,
         globalConfig: { baseUrl: authUrl, appUrl },
         config: providersConfig.getConfig(providerId),
         logger,