+++ plugins/catalog-backend/src/ingestion/processors/BuiltinKindsEntityProcessor.ts
@@ -15,15 +15,33 @@
  */
 
 import {
+  ApiEntity,
   apiEntityV1alpha1Validator,
+  ComponentEntity,
   componentEntityV1alpha1Validator,
   Entity,
+  getEntityName,
+  GroupEntity,
   groupEntityV1alpha1Validator,
   locationEntityV1alpha1Validator,
+  LocationSpec,
+  parseEntityRef,
+  RELATION_API_CONSUMED_BY,
+  RELATION_API_PROVIDED_BY,
+  RELATION_CHILD_OF,
+  RELATION_CONSUMES_API,
+  RELATION_HAS_MEMBER,
+  RELATION_MEMBER_OF,
+  RELATION_OWNED_BY,
+  RELATION_OWNER_OF,
+  RELATION_PARENT_OF,
+  RELATION_PROVIDES_API,
   templateEntityV1alpha1Validator,
+  UserEntity,
   userEntityV1alpha1Validator,
 } from '@backstage/catalog-model';
-import { CatalogProcessor } from './types';
+import * as result from './results';
+import { CatalogProcessor, CatalogProcessorEmit } from './types';
 
 export class BuiltinKindsEntityProcessor implements CatalogProcessor {
   private readonly validators = [
@@ -35,6 +53,25 @@ export class BuiltinKindsEntityProcessor implements CatalogProcessor {
     userEntityV1alpha1Validator,
   ];
 
+  async preProcessEntity(entity: Entity): Promise<Entity> {
+    // NOTE(freben): Part of Group field deprecation on Nov 22nd, 2020. Fields
+    // scheduled for removal Dec 6th, 2020. This code can be deleted after that
+    // point. See https://github.com/backstage/backstage/issues/3049
+    if (
+      entity.apiVersion === 'backstage.io/v1alpha1' &&
+      entity.kind === 'Group' &&
+      entity.spec
+    ) {
+      if (!entity.spec.ancestors) {
+        entity.spec.ancestors = [];
+      }
+      if (!entity.spec.descendants) {
+        entity.spec.descendants = [];
+      }
+    }
+    return entity;
+  }
+
   async validateEntityKind(entity: Entity): Promise<boolean> {
     for (const validator of this.validators) {
       const result = await validator.check(entity);
@@ -44,5 +81,127 @@ export class BuiltinKindsEntityProcessor implements CatalogProcessor {
     }
 
     return false;
+  }
+
+  async postProcessEntity(
+    entity: Entity,
+    _location: LocationSpec,
+    emit: CatalogProcessorEmit,
+  ): Promise<Entity> {
+    const selfRef = getEntityName(entity);
+
+    /*
+     * Utilities
+     */
+
+    function doEmit(
+      targets: string | string[] | undefined,
+      context: { defaultKind: string; defaultNamespace: string },
+      outgoingRelation: string,
+      incomingRelation: string,
+    ): void {
+      if (!targets) {
+        return;
+      }
+      for (const target of [targets].flat()) {
+        const targetRef = parseEntityRef(target, context);
+        emit(
+          result.relation({
+            source: selfRef,
+            type: outgoingRelation,
+            target: targetRef,
+          }),
+        );
+        emit(
+          result.relation({
+            source: targetRef,
+            type: incomingRelation,
+            target: selfRef,
+          }),
+        );
+      }
+    }
+
+    /*
+     * Emit relations for the Component kind
+     */
+
+    if (entity.kind === 'Component') {
+      const component = entity as ComponentEntity;
+      doEmit(
+        component.spec.owner,
+        { defaultKind: 'Group', defaultNamespace: selfRef.namespace },
+        RELATION_OWNED_BY,
+        RELATION_OWNER_OF,
+      );
+      doEmit(
+        component.spec.implementsApis,
+        { defaultKind: 'API', defaultNamespace: selfRef.namespace },
+        RELATION_PROVIDES_API,
+        RELATION_API_PROVIDED_BY,
+      );
+      doEmit(
+        component.spec.providesApis,
+        { defaultKind: 'API', defaultNamespace: selfRef.namespace },
+        RELATION_PROVIDES_API,
+        RELATION_API_PROVIDED_BY,
+      );
+      doEmit(
+        component.spec.consumesApis,
+        { defaultKind: 'API', defaultNamespace: selfRef.namespace },
+        RELATION_CONSUMES_API,
+        RELATION_API_CONSUMED_BY,
+      );
+    }
+
+    /*
+     * Emit relations for the API kind
+     */
+
+    if (entity.kind === 'API') {
+      const api = entity as ApiEntity;
+      doEmit(
+        api.spec.owner,
+        { defaultKind: 'Group', defaultNamespace: selfRef.namespace },
+        RELATION_OWNED_BY,
+        RELATION_OWNER_OF,
+      );
+    }
+
+    /*
+     * Emit relations for the User kind
+     */
+
+    if (entity.kind === 'User') {
+      const user = entity as UserEntity;
+      doEmit(
+        user.spec.memberOf,
+        { defaultKind: 'Group', defaultNamespace: selfRef.namespace },
+        RELATION_MEMBER_OF,
+        RELATION_HAS_MEMBER,
+      );
+    }
+
+    /*
+     * Emit relations for the Group kind
+     */
+
+    if (entity.kind === 'Group') {
+      const group = entity as GroupEntity;
+      doEmit(
+        group.spec.parent,
+        { defaultKind: 'Group', defaultNamespace: selfRef.namespace },
+        RELATION_CHILD_OF,
+        RELATION_PARENT_OF,
+      );
+      doEmit(
+        group.spec.children,
+        { defaultKind: 'Group', defaultNamespace: selfRef.namespace },
+        RELATION_PARENT_OF,
+        RELATION_CHILD_OF,
+      );
+    }
+
+    return entity;
   }
 }