+++ plugins/kubernetes/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-kubernetes
 
+## 0.3.0
+
+### Minor Changes
+
+- 1166fcc36: add kubernetes selector to component model
+
+### Patch Changes
+
+- Updated dependencies [475fc0aaa]
+- Updated dependencies [1166fcc36]
+- Updated dependencies [1185919f3]
+  - @backstage/core@0.3.2
+  - @backstage/catalog-model@0.3.0
+  - @backstage/plugin-kubernetes-backend@0.2.0
+
 ## 0.2.1
 
 ### Patch Changes