+++ plugins/sonarqube/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-sonarqube
 
+## 0.1.4
+
+### Patch Changes
+
+- 26484d413: Add configuration schema
+- Updated dependencies [475fc0aaa]
+- Updated dependencies [1166fcc36]
+- Updated dependencies [1185919f3]
+  - @backstage/core@0.3.2
+  - @backstage/catalog-model@0.3.0
+
 ## 0.1.3
 
 ### Patch Changes